+++ src/TestInfrastructure/Framework/ConfigurableVsShell.cs
@@ -79,7 +79,7 @@ namespace SonarLint.VisualStudio.Integration.UnitTests
             throw new NotImplementedException();
         }
 
-        int IVsShell.LoadUILibrary(ref Guid guidPackage, uint dwExFlags, out uint phinstOut)
+        int IVsShell.LoadUILibrary(ref Guid guidPackage, uint dwExFlags, out IntPtr phinstOut)
         {
             throw new NotImplementedException();
         }
